+++ app/routes/problem/$pid/submit.tsx
@@ -3,27 +3,24 @@ import { db } from "~/utils/db.server";
 import { s3 } from "~/utils/s3.server";
 import { invariant } from "~/utils/invariant";
 import { codeScheme, idScheme } from "~/utils/scheme";
-import { findSessionUid } from "~/utils/sessions";
 import { judge } from "~/utils/judge.server";
+import { guaranteePermission, Permissions } from "~/utils/permission";
 
-export const loader: LoaderFunction = async ({ request }) => {
-  const self = await findSessionUid(request);
+export const loader: LoaderFunction = async ({ request, params }) => {
+  const pid = invariant(idScheme.safeParse(params.pid), { status: 404 });
 
-  if (!self) {
-    return redirect(`/login?redirect=${new URL(request.url).pathname}`);
-  }
+  await guaranteePermission(request, Permissions.Problem.Submit, { pid });
 
   return null;
 };
 
-export const action: ActionFunction = async ({ params, request }) => {
-  const uid = await findSessionUid(request);
-
-  if (!uid) {
-    return redirect(`/login?redirect=${new URL(request.url).pathname}`);
-  }
-
+export const action: ActionFunction = async ({ request, params }) => {
   const pid = invariant(idScheme.safeParse(params.pid), { status: 404 });
+
+  // 检查提交权限，用户必定已经登录
+  const uid = (await guaranteePermission(request, Permissions.Problem.Submit, {
+    pid,
+  }))!;
 
   const form = await request.formData();
   const code = invariant(codeScheme.safeParse(form.get("code")));